feat: move types closer to their owners
