+++ src/modules/course-list/@types/types.ts
@@ -1,0 +1,6 @@
+import { ICourse, ICourseMeta } from '../../../shared/@types/types';
+
+export type CourseItemPreview = Pick<
+  ICourse,
+  'id' | 'title' | 'previewImageLink' | 'lessonsCount' | 'rating'
+> & { meta: Pick<ICourseMeta, 'skills' | 'courseVideoPreview'> };
